fetcher/requirements.txt
python-dotenv
PyYAML==5.3.1
requests
Flask
gunicorn
paho-mqtt==1.6.1
